5000: divide constants and functions for old trigger form
