+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/TriggerFormAssetWallet.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import FormikSelect from '../../../../../components/formik-santiment-ui/FormikSelect'
 import FormikInput from '../../../../../components/formik-santiment-ui/FormikInput'
-import { ASSETS_FILTERS, ETH_WALLET_METRIC } from '../../../utils/utils'
+import { ASSETS_FILTERS, ETH_WALLET_METRIC } from '../../../utils/constants'
 import styles from '../signal/TriggerForm.module.scss'
 
 const propTypes = {
